Refaktor. Slå sammen VilkårService og VilkårStegService.
