+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/stønadsvilkår/VilkårController.kt
@@ -25,7 +25,6 @@ import java.util.UUID
 @Validated
 class VilkårController(
     private val vilkårService: VilkårService,
-    private val vilkårStegService: VilkårStegService,
     private val tilgangService: TilgangService,
     // private val gjenbrukVilkårService: GjenbrukVilkårService,
 ) {
@@ -42,7 +41,7 @@ class VilkårController(
         tilgangService.validerTilgangTilBehandling(svarPåVilkårDto.behandlingId, AuditLoggerEvent.UPDATE)
         tilgangService.validerHarSaksbehandlerrolle()
         try {
-            return vilkårStegService.oppdaterVilkår(svarPåVilkårDto)
+            return vilkårService.oppdaterVilkår(svarPåVilkårDto)
         } catch (e: Exception) {
             val delvilkårJson = objectMapper.writeValueAsString(svarPåVilkårDto.delvilkårsett)
             secureLogger.warn(
@@ -58,14 +57,14 @@ class VilkårController(
     fun nullstillVilkår(@RequestBody request: OppdaterVilkårDto): VilkårDto {
         tilgangService.validerTilgangTilBehandling(request.behandlingId, AuditLoggerEvent.DELETE)
         tilgangService.validerHarSaksbehandlerrolle()
-        return vilkårStegService.nullstillVilkår(request)
+        return vilkårService.nullstillVilkår(request)
     }
 
     @PostMapping("ikkevurder")
     fun settVilkårTilSkalIkkeVurderes(@RequestBody request: OppdaterVilkårDto): VilkårDto {
         tilgangService.validerTilgangTilBehandling(request.behandlingId, AuditLoggerEvent.UPDATE)
         tilgangService.validerHarSaksbehandlerrolle()
-        return vilkårStegService.settVilkårTilSkalIkkeVurderes(request)
+        return vilkårService.settVilkårTilSkalIkkeVurderes(request)
     }
 
     @GetMapping("{behandlingId}")
